refactor: restructure profile form to match project patterns
  - Extract profile schema to separate file with Zod validation
  - Create useProfileForm hook for form logic and state management
  - Create ProfileForm component for UI rendering
  - Simplify Profile page to wrapper component
  - Remove useEffect in favor of RHF's values prop for automatic updates
  - Maintain separation of concerns matching login/register pattern
